+++ src/tests/test_integration.py
@@ -251,6 +251,32 @@ def test_config_file(env):
     assert 'D103' not in err
 
 
+def test_sectionless_config_file(env):
+    """Test that config files without a valid section name issue a warning."""
+    with env.open('config.ini', 'wt') as conf:
+        conf.write('[pdcstl]')
+        config_path = conf.name
+
+    _, err, code = env.invoke('--config={}'.format(config_path))
+    assert code == 0
+    assert 'Configuration file does not contain a pydocstyle section' in err
+
+    with env.open('example.py', 'wt') as example:
+        example.write(textwrap.dedent("""\
+            def foo():
+                pass
+        """))
+
+    with env.open('tox.ini', 'wt') as conf:
+        conf.write('[pdcstl]\n')
+        conf.write('ignore = D100')
+
+    out, err, code = env.invoke()
+    assert code == 1
+    assert 'D100' in out
+    assert 'file does not contain a pydocstyle section' not in err
+
+
 def test_config_path(env):
     """Test that options are correctly loaded from a specific config file.